SDT-201: Updates to classes used to convert XML to Json
Second set of updates to classes used to convert a CBUS XML request to Json for passing on to CMC.

Changed datatypes of following fields to Long:
- ClaimRequest claimAmount and solicitorCost fields
- Interest dailyAmount and claimAmountInterestBase fields
- JudgmentRequest interest field

Datatype needed to match that of corresponding field in XML.

Also includes update to Judgement.xml test data file which was missed from previous commit.

+++ consumers/src/main/java/uk/gov/moj/sdt/cmc/consumers/request/claim/ClaimRequest.java
@@ -69,10 +69,10 @@ public class ClaimRequest {
     private Interest interest;
 
     @Getter
-    private Double claimAmount;
+    private Long claimAmount;
 
     @Getter
-    private Double solicitorCost;
+    private Long solicitorCost;
 
     @JsonProperty(value = "particularsLines")
     @Getter
@@ -80,7 +80,7 @@ public class ClaimRequest {
 
     private SotSignature sotSignature;
 
-    public Double getInterestDailyAmount() {
+    public Long getInterestDailyAmount() {
         return interest == null ? null : interest.getDailyAmount();
     }
 
@@ -94,7 +94,7 @@ public class ClaimRequest {
         return interest == null ? null : interest.getClaimDate();
     }
 
-    public Double getClaimAmountInterestBase() {
+    public Long getClaimAmountInterestBase() {
         return interest == null ? null : interest.getClaimAmountInterestBase();
     }
